+++ uamqp/__init__.py
@@ -39,17 +39,17 @@ def send_message(target, data, auth=None, debug=False):
     """Send a single message to AMQP endpoint.
 
     :param target: The target AMQP endpoint.
-    :type target: str, bytes or ~uamqp.Target
+    :type target: str, bytes or ~uamqp.address.Target
     :param data: The contents of the message to send.
-    :type data: str, bytes or ~uamqp.Message
+    :type data: str, bytes or ~uamqp.message.Message
     :param auth: The authentication credentials for the endpoint.
-     This should be one of the subclasses of ~uamqp.AMQPAuth. Currently
+     This should be one of the subclasses of ~uamqp.authentication.AMQPAuth. Currently
      this includes:
-        ~uamqp.authentication.SASLAnonymous
-        ~uamqp.authentication.SASLPlain
-        ~uamqp.authentication.SASTokenAuth
+        - ~uamqp.authentication.SASLAnonymous
+        - ~uamqp.authentication.SASLPlain
+        - ~uamqp.authentication.SASTokenAuth
      If no authentication is supplied, SASLAnnoymous will be used by default.
-    :type auth: ~uamqp.AMQPAuth
+    :type auth: ~uamqp.authentication.AMQPAuth
     :param debug: Whether to turn on network trace logs. If `True`, trace logs
      will be logged at INFO level. Default is `False`.
     :type debug: bool
@@ -65,22 +65,22 @@ def receive_message(source, auth=None, timeout=0, debug=False):
     """Receive a single message from an AMQP endpoint.
 
     :param source: The AMQP source endpoint to receive from.
-    :type source: str, bytes or ~uamqp.Source
+    :type source: str, bytes or ~uamqp.address.Source
     :param auth: The authentication credentials for the endpoint.
-     This should be one of the subclasses of ~uamqp.AMQPAuth. Currently
+     This should be one of the subclasses of ~uamqp.authentication.AMQPAuth. Currently
      this includes:
-        ~uamqp.authentication.SASLAnonymous
-        ~uamqp.authentication.SASLPlain
-        ~uamqp.authentication.SASTokenAuth
+        - ~uamqp.authentication.SASLAnonymous
+        - ~uamqp.authentication.SASLPlain
+        - ~uamqp.authentication.SASTokenAuth
      If no authentication is supplied, SASLAnnoymous will be used by default.
-    :type auth: ~uamqp.AMQPAuth
+    :type auth: ~uamqp.authentication.AMQPAuth
     :param timeout: The timeout in seconds after which to return None if no messages
      are retrieved. If set to `0` (the default), the receiver will not timeout and
      will continue to wait for messages until interrupted.
     :param debug: Whether to turn on network trace logs. If `True`, trace logs
      will be logged at INFO level. Default is `False`.
     :type debug: bool
-    :rtype: ~uamqp.Message or None
+    :rtype: ~uamqp.message.Message or None
     """
     received = receive_messages(source, auth=auth, max_batch_size=1, timeout=timeout, debug=debug)
     if received:
@@ -92,15 +92,15 @@ def receive_messages(source, auth=None, max_batch_size=None, timeout=0, debug=Fa
     """Receive a batch of messages from an AMQP endpoint.
 
     :param source: The AMQP source endpoint to receive from.
-    :type source: str, bytes or ~uamqp.Source
+    :type source: str, bytes or ~uamqp.address.Source
     :param auth: The authentication credentials for the endpoint.
-     This should be one of the subclasses of ~uamqp.AMQPAuth. Currently
+     This should be one of the subclasses of ~uamqp.authentication.AMQPAuth. Currently
      this includes:
-        ~uamqp.authentication.SASLAnonymous
-        ~uamqp.authentication.SASLPlain
-        ~uamqp.authentication.SASTokenAuth
+        - ~uamqp.authentication.SASLAnonymous
+        - ~uamqp.authentication.SASLPlain
+        - ~uamqp.authentication.SASTokenAuth
      If no authentication is supplied, SASLAnnoymous will be used by default.
-    :type auth: ~uamqp.AMQPAuth
+    :type auth: ~uamqp.authentication.AMQPAuth
     :param max_batch_size: The maximum number of messages to return in a batch. If the
      receiver receives a smaller number than this, it will not wait to return them so
      the actual number returned can be anything up to this value. If the receiver reaches
@@ -111,7 +111,7 @@ def receive_messages(source, auth=None, max_batch_size=None, timeout=0, debug=Fa
     :param debug: Whether to turn on network trace logs. If `True`, trace logs
      will be logged at INFO level. Default is `False`.
     :type debug: bool
-    :rtype: list[~uamqp.Message]
+    :rtype: list[~uamqp.message.Message]
     """
     if max_batch_size:
         kwargs['prefetch'] = max_batch_size
@@ -157,6 +157,7 @@ class _Platform:
 
 def get_platform_info():
     """Gets the current platform information.
+
     :rtype: str
     """
     return str(c_uamqp.get_info())