+++ uamqp/async/mgmt_operation_async.py
@@ -25,7 +25,7 @@ class MgmtOperationAsync(MgmtOperation):
 
     :param session: The AMQP session to use for the operation. Nee send and
      receive links will be created in this Session.
-    :type session: ~uamqp.SessionAsync
+    :type session: ~uamqp.async.session_async.SessionAsync
     :param target: The AMQP node to send the request to.
      The default is `b"$management"`
     :type target: bytes or str
@@ -71,11 +71,11 @@ class MgmtOperationAsync(MgmtOperation):
          an application property on the message.
         :type op_type: bytes
         :param message: The message to send in the management request.
-        :type message: ~uamqp.Message
+        :type message: ~uamqp.message.Message
         :param timeout: Provide an optional timeout in milliseconds within which a response
          to the management request must be received.
         :type timeout: int
-        :rtype: ~uamqp.Message
+        :rtype: ~uamqp.message.Message
         """
         start_time = self._counter.get_current_ms()
         operation_id = str(uuid.uuid4())