+++ uamqp/client.py
@@ -174,7 +174,7 @@ class AMQPClient:
 
         :param connection: An existing Connection that may be shared between
          multiple clients.
-        :type connetion: ~uamqp.Connection
+        :type connetion: ~uamqp.connection.Connection
         """
         # pylint: disable=protected-access
         if self._session:
@@ -241,7 +241,7 @@ class AMQPClient:
         and the available options will depend on the target service.
 
         :param message: The message to send in the management request.
-        :type message: ~uamqp.Message
+        :type message: ~uamqp.message.Message
         :param operation: The type of operation to be performed. This value will
          be service-specific, but common values incluse READ, CREATE and UPDATE.
          This value will be added as an application property on the message.
@@ -263,7 +263,7 @@ class AMQPClient:
          response body which can vary between services due to the spec still being in draft.
          The default is `b"statusDescription"`.
         :type description_fields: bytes
-        :rtype: ~uamqp.Message
+        :rtype: ~uamqp.message.Message
         """
         timeout = False
         auth_in_progress = False
@@ -320,8 +320,8 @@ class SendClient(AMQPClient):
     """An AMQP client for sending messages.
 
     :param target: The target AMQP service endpoint. This can either be the URI as
-     a string or a ~uamqp.Target object.
-    :type target: str, bytes or ~uamqp.Target
+     a string or a ~uamqp.address.Target object.
+    :type target: str, bytes or ~uamqp.address.Target
     :param auth: Authentication for the connection. If none is provided SASL Annoymous
      authentication will be used.
     :type auth: ~uamqp.authentication.AMQPAuth
@@ -487,9 +487,9 @@ class SendClient(AMQPClient):
         to the queue.
 
         :param messages: A message to send. This can either be a single instance
-         of ~uamqp.Message, or multiple messages wrapped in an instance
-         of ~uamqp.BatchMessage.
-        :type message: ~uamqp.Message
+         of ~uamqp.message.Message, or multiple messages wrapped in an instance
+         of ~uamqp.message.BatchMessage.
+        :type message: ~uamqp.message.Message
         """
         for message in messages.gather():
             message.idle_time = self._counter.get_current_ms()
@@ -499,9 +499,9 @@ class SendClient(AMQPClient):
         """Send a single message or batched message.
 
         :param messages: A message to send. This can either be a single instance
-         of ~uamqp.Message, or multiple messages wrapped in an instance
-         of ~uamqp.BatchMessage.
-        :type message: ~uamqp.Message
+         of ~uamqp.message.Message, or multiple messages wrapped in an instance
+         of ~uamqp.message.BatchMessage.
+        :type message: ~uamqp.message.Message
         :param close_on_done: Close the client once the message is sent. Default is `False`.
         :type close_on_done: bool
         :raises: ~uamqp.errors.MessageSendFailed if message fails to send after retry policy
@@ -571,8 +571,8 @@ class ReceiveClient(AMQPClient):
     """An AMQP client for receiving messages.
 
     :param target: The source AMQP service endpoint. This can either be the URI as
-     a string or a ~uamqp.Source object.
-    :type target: str, bytes or ~uamqp.Source
+     a string or a ~uamqp.address.Source object.
+    :type target: str, bytes or ~uamqp.address.Source
     :param auth: Authentication for the connection. If none is provided SASL Annoymous
      authentication will be used.
     :type auth: ~uamqp.authentication.AMQPAuth
@@ -709,7 +709,7 @@ class ReceiveClient(AMQPClient):
     def _message_generator(self):
         """Iterate over processed messages in the receive queue.
 
-        :rtype: generator[~uamqp.Message]
+        :rtype: generator[~uamqp.message.Message]
         """
         self.open()
         auto_complete = self.auto_complete
@@ -741,7 +741,8 @@ class ReceiveClient(AMQPClient):
         Additionally if the client is retrieving messages for a batch
         or iterator, the message will be added to an internal queue.
 
-        :param message: ~c_uamqp.Message
+        :param message: Received message.
+        :type message: ~uamqp.message.Message
         """
         self._was_message_received = True
         if self._message_received_callback:
@@ -772,8 +773,8 @@ class ReceiveClient(AMQPClient):
          the prefetch value will be used.
         :type max_batch_size: int
         :param on_message_received: A callback to process messages as they arrive from the
-         service. It takes a single argument, a ~uamqp.Message object.
-        :type on_message_received: callable[~uamqp.Message]
+         service. It takes a single argument, a ~uamqp.message.Message object.
+        :type on_message_received: callable[~uamqp.message.Message]
         :param timeout: I timeout in milliseconds for which to wait to receive any messages.
          If no messages are received in this time, an empty list will be returned. If set to
          0, the client will continue to wait until at least one message is received. The
@@ -826,8 +827,8 @@ class ReceiveClient(AMQPClient):
         to be explicitly settled during the callback, otherwise it will be released.
 
         :param on_message_received: A callback to process messages as they arrive from the
-         service. It takes a single argument, a ~uamqp.Message object.
-        :type on_message_received: callable[~uamqp.Message]
+         service. It takes a single argument, a ~uamqp.message.Message object.
+        :type on_message_received: callable[~uamqp.message.Message]
         """
         self.open()
         self._message_received_callback = on_message_received
@@ -848,8 +849,8 @@ class ReceiveClient(AMQPClient):
         criteria, pass in a callback.
 
         :param on_message_received: A callback to process messages as they arrive from the
-         service. It takes a single argument, a ~uamqp.Message object.
-        :type on_message_received: callable[~uamqp.Message]
+         service. It takes a single argument, a ~uamqp.message.Message object.
+        :type on_message_received: callable[~uamqp.message.Message]
         """
         self._message_received_callback = on_message_received
         self._received_messages = queue.Queue()